+++ tests/dummy/app/controllers/components-acceptance-tests/flexberry-objectlistview/custom-filter.js
@@ -6,6 +6,8 @@ export default ListFormController.extend({
 
   filterByAllWords: false,
 
+  filterProjectionName: undefined,
+
   /**
     Cout of list loading.