+++ tests/dummy/app/models/ember-flexberry-dummy-suggestion.js
@@ -378,4 +378,15 @@ Model.defineProjection('SuggestionEWithKarma', 'ember-flexberry-dummy-suggestion
   })
 });
 
+// Projection for test on filter on folv toolbar.
+Model.defineProjection('TestFilterOnToolbarView', 'ember-flexberry-dummy-suggestion', {
+  type: belongsTo('ember-flexberry-dummy-suggestion-type', 'Type', {
+    name: attr('Name', {
+      hidden: true
+    })
+  }, {
+    displayMemberPath: 'name'
+  })
+});
+
 export default Model;